Moved the route that updates the weights when a user connects to the connections router. Iniitally had it in user-actions cause i thought it was like its own thing but it does just go under connections
